Still not working, trying some new things with Farshad

+++ Mantarray/Inc/Bus.h
@@ -32,7 +32,7 @@ typedef struct
 }InternalBus_t;
 
 InternalBus_t * internal_bus_create(GPIO_TypeDef *bus_line,uint16_t bus_pins,GPIO_TypeDef *cl_bus,uint16_t cl_pin,GPIO_TypeDef *ak_bus,uint16_t ak_pin);
-void internal_bus_write_data_frame(InternalBus_t *thisInternalBus,uint32_t *buffer,uint8_t buffer_len);
+void internal_bus_write_data_frame(InternalBus_t *thisInternalBus,uint8_t *buffer,uint8_t buffer_len);
 void internal_bus_utilize(InternalBus_t *thisInternalBus);
 void internal_bus_release(InternalBus_t *thisInternalBus);
